refactor: split backlog models into their own file
for my own sanity
